Refactor getModelName function to import from "src/ts/model/names"
